[keyserver] reorganize rust-node-addon
Summary:
before I put up the diffs for login_user, I wanted to do a little bit of reorganizing. now, delete_user and register_user are submodules of the identity_client module, which is more intuitive IMO.

I also added an interceptor in the delete_user fn so that we can authorize delete requests to the Identity service

There is some repeated code, but I didn't want to spend more time figuring out how to de-duplicate it. Updated the following task to track this work: https://linear.app/comm/issue/ENG-2733/refactor-native-rust-library-and-rust-node-addon-client-logic

Test Plan: tested the DeleteUser RPC with the new interceptor, things worked as expected

Reviewers: jon, bartek, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6913