[keyserver] add loginUserPake function to rust-node-addon, call it from loginResponder in js
Summary:
Added loginUserPake based on the existing, equivalent function in our native_rust_library. The only major
difference is the error types.

I haven't tested this fully yet, but need help with the `sessionInitializationInfo` type in flow. HashMaps in rust
map to Objects in JavaScript with napi-rs, but I think we usually have stronger typing than that in flow.

Test Plan: tbd

Reviewers: ashoat

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D6914