[keyserver] add loginUserPake function to rust-node-addon, call it from loginResponder in js
Summary:
Added loginUserPake based on the existing, equivalent function in our native_rust_library. The only major
difference is the error types.

we call it in loginResponder here. will add it to siweAuthResponder the diff after the next

Test Plan: this is tested along with registerUser, see the test plan in the subsequent diff

Reviewers: ashoat, jon, bartek, atul

Reviewed By: ashoat

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D6914