[lib] Include fact that you were tagged in sidebar creation message
Summary:
In the stack ending with D6874, I made it so that if you were @-mentioned in the OP, you would get a notif that ended with "that tagged you". (Part 1 from [this Linear comment](https://linear.app/comm/issue/ENG-3042#comment-b78d044f).)

This diff completes the stack that makes it so if you were @-mentioned in the message that triggers the creation of the sidebar, you will get a notif that ends with "and tagged you". (Part 2 from [this Linear comment](https://linear.app/comm/issue/ENG-3042#comment-b78d044f).)

Depends on D6937

Test Plan:
Flow, and tested in combination with the next diffs to confirm that getting @-tagged in a message that triggers the creation of a sidebar will update the `SIDEBAR_CREATION` notif
to be suffixed with "and tagged you"

Reviewers: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D6938