[web] Restyle search area in FilterPanel
Summary:
Issue: https://linear.app/comm/issue/ENG-3014/add-a-button-to-close-the-filter-panel
There is supposed to be a button for closing the filter panel, inside of the filter panel. Additionally some styles around it were changed, so I changed them as well to keep the panel looking consistent. Note that in the issues comments it had been decided to keep the search bars background colour consistent with the rest of the app for now.
The filters buttnon (the one used for opening) is supposed to disappear when the panel is open

{F384736}

{F384781}

Test Plan: run web app, checked that items in the filter panel display correctly.

Reviewers: michal, kamil, kuba

Reviewed By: michal

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6752