[web] Restyle search area in FilterPanel
Summary:
Issue: https://linear.app/comm/issue/ENG-3014/add-a-button-to-close-the-filter-panel
There is supposed to be a buttoon for closing the filter panel, inside of the filter panel. Additionally some styles aroud it were changed, so I changed them as well to keep the panel looking consistent. Note that in the issues
comments it had been decided to keep the search bars background colour consistent with the rest of the app for now.
Also note that for now
- the button doesn't do anything
- the filters button is still present when the filter panel is open
This will be changed in future diffs

Test Plan: run web app, checked that items in the filter panel display correctly.

Reviewers: michal, kamil, kuba

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6752