[web] Restyle filters button
Summary:
Issue: https://linear.app/comm/issue/ENG-2964/restyle-the-filters-button
Restyling the filters button. Note that in the comments, it has been decided that we'll use a different icon than in  the designs.
For now keeping the button after opening the filter panel. This will be changed in the future

{F382519}

Test Plan: Run web app, checked that the button displays correctly, and pressing it opens the filter panel

Reviewers: michal, kamil, kuba

Reviewed By: kuba

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6740