Prepare to conditionally process rescinds in NSE.
Summary:
This differential is intended to provide reasonable test plans for further differentials in this stack. On its own it is a no-op, but if we remove `codeVersion > 1000` condition then both visible notifications and rescinds start to be processed by NSE. This way we have a simple
method to test further differentials where we actually do some notifications processing in NSE and we are not introducing any breaking changes in the middle of the stack. Final differential in the stack will remove this condition and replace it with particular release build number.

What this differential does is to structure rescind as if it was a regular notification. This is required to launch NotificationService to process it. Using notifications filtering capability we make sure this artificial content is not displayed to the user (matter of next differential).

Test Plan:
Remove `codeVersion > 1000` condition. Visible notifications will start to be persisted in temporary flat file in NSE. Send some notifications, and watch in the debugger persisted notifications payloads in `moveMessagesToDatabase` method of AppDelegate. Send a rescind. It
won't work but you will see a strange notification with `PLACEHOLDER` as body.

Reviewers: bartek, atul, tomek, kamil

Reviewed By: atul

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6917