Silence rescinds in NSE
Summary: This differential implements code to silence rescinds so that they are not visible to users. No actual rescind processing is done.

Test Plan:
Remove `codeVersion > 1000` from `send.js` and `rescind.js` files in keyserver in `prepareIOSNotification` functions. Generate rescind. Ensure it is not displayed to the user. This is how we use notifications filtering capability:
https://developer.apple.com/documentation/bundleresources/entitlements/com_apple_developer_usernotifications_filtering?language=objc

Reviewers: bartek, atul, tomek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6918