Use references of DatabaseClient
Summary: DatabaseClient is essentially an instance factory, we
don't need to be taking ownership of it on every function call;
thus forcing a `.clone()` on the caller.

Test Plan:
This will be tested more thoroughly in a later diff, however,
the changes should just reduce the number of deep copies we do
of the "instance factory".
```
cd services/identity
cargo build
```

Reviewers: varun, max, bartek

Subscribers: