[native] Create feature flags context
Summary:
This context will be responsible for fetching and providing a configuration. At this point it just returns the default one.
Initially I tried to introduce this context in lib, so that it can be used by web, but that's not trivial, as this context requires `staff` flag that we currently compute on native and moving it to web also requires some thought as we e.g. hardcode `isStaffRelease` flag which doesn't make sense on web. Also, the initial release of feature flags service covers mobile only.

Depends on D6947

Test Plan: Consume the context in the app and console log it.

Reviewers: kamil, bartek, inka

Reviewed By: bartek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6948