[native] Fetch and parse feature flags response
Summary:
Fetch a config, parse, and use in a state.

Depends on D6948

Test Plan: The service isn't yet deployed using https, so the networking stuff wasn't tested. I've hardcoded a response and verified that setting the state works ok. After https is in place, I'm going to update this test plan.

Reviewers: kamil, bartek, inka