[keyserver] Introduce a new thread permission for pinning / unpinning messages
Summary:
We need to introduce s new thread permission that limits who is able to manage the pins for a thread. At the moment, this is only granted to Admins, so there's a couple of things that need to be done
here:
1. Update the baseAdminPriviledges so new chats with admins will have the correct, updated permissions
2. Run a migration to cover existing threads
3. Include the necessary migration to create the new `pinned_messages` table from the previous diff.

The general overview is to update the `roles` table for Admins, and then call `updateRolesAndPermissionsForAllThreads()` which should handle updating the permissions for each of the threads given the new
roles.

The migration was discussed in the [[ https://linear.app/comm/issue/ENG-3181/introduce-a-new-thread-permission-for-pinning-unpinning-messages | Linear ]] task.

Depends on D6924

Test Plan:
1. Confirm that the migration is successful

2. Confirm that the `roles` table is updated for Admins (expected result is that the query only finds Admins to have the new permissions)

3. Confirm that the `memberships` table for each of the threads is updated (expected result is that the query finds the permissions to be NOT NULL in each thread for user 256, the admin user locally)

Reviewers: atul, ginsu, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6928