[keyserver] Add a function to toggle the pin status of a message
Summary: Create a function that toggles the pin status of a message provided the request of the type `ToggleMessagePinRequest`. We check the permissions for the requester to ensure they have the correct permissions to manage message pins.

Linear: https://linear.app/comm/issue/ENG-3184/add-a-function-to-toggle-the-pin-status-of-a-message

Depends on D6928

(Note: I put this in `thread-updaters` since I plan to return some kind of robotext and it seems like convention that it's handled here (i.e. removeMembers, joinThread, etc). Feel free to correct me if I've
not scoped the conventions correctly though).

Test Plan: The test plan for this function will be combined into the test plan for the subsequent diff, since the endpoints and responders will be set up there.

Reviewers: atul, ginsu, tomek

Subscribers: