[keyserver/lib] Set up responders, endpoints, and related functions to toggle a message pin
Summary: Now that the function to `INSERT` and `DELETE` from MariaDB is set up, this diff handles setting up the endpoint, responder and validator, as well as the action method that'll be called via
`useServerCall` client-side.

Linear: https://linear.app/comm/issue/ENG-3187/set-up-responders-endpoints-and-related-functions-to-toggle-a-message

Depends on D6929

Test Plan: The testing for both the previous diff and this diff are combined here. I tested the stack so far by doing the following:

1. Finding a place to call `toggleMessagePin` on the client. For this, I used the thread settings media gallery on `web`.

2. Log into an admin account locally.

3. Call `toggleMessagePin` with the following parameters: `{messageID: "123", threadID: "83895", isPinned: false}`. I confirm that the insertion into `pinned_messages` was successful.

4. Call `toggleMessagePin` with the following parameters: `{messageID: "123", threadID: "83895", isPinned: false}`. I confirm that the deletion from `pinned_messages` was successful.

5. Log into a non-admin account part of thread 83895

6. Call `toggleMessagePin` with the same parameters, and confirm that the `ServerError` throws

Reviewers: atul, ginsu, tomek

Subscribers:

Differential Revision: https://phab.comm.dev/D6930