[identity] change siwe signature type
Summary: in the keyserver, the siwe signature is a string. we were previously handling it as a byte vector in the identity service, so this change is to be consistent with the keyserver.

Depends on D6941

Test Plan: will be tested in next diff where we use the WalletLoginRequest protobuf message type

Reviewers: jon, ashoat, bartek

Subscribers: