[lib] Add 'hasBeenEdited' status to messages
Summary: Added 'hasBeenEdited' variable to track if the message was edited. Will be used to displaying 'edited' label.

Test Plan: Checked if the app compiles and works. Then tested if the status is set properly by changing labels of tooltip buttons to the edited status (instead of 'Reply' i displayed edited status).

Reviewers: michal, inka

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6966