[web] Add to redux state picked community in the Chat tab
Summary:
We will need the information about the selected community in the Chat drawer to be accesable from different parts of the code. For this reason we will keep it in the redux state, just like we
already keep the selected community for the claendar tab in redux.
isse: https://linear.app/comm/issue/ENG-3176/make-pressing-a-drawer-item-in-chat-tab-filter-available-chats

Test Plan:
Rub web app, checked that selecting a community in calendar still works
Reviewers:
kamil, tomek, michal

Subscribers: