[web] Add handlers for drawer community items
Summary:
Adding hanndlers that will be used by the community items in the drawer
They will be used like `CommunityDrawerItemHandler` were used before in `CommunityDrawerItem`. They provide drawer community items with an `onClick` function, that defines how the items is
supposed to behave, an `expanded` boolean, that tells the community whether it is expanded, and a `isActive` boolean that tells whether the user has currently that exact chat selected.
issue: https://linear.app/comm/issue/ENG-3176/make-pressing-a-drawer-item-in-chat-tab-filter-available-chats

Test Plan:
Tested with later diffs

Reviewers:
kamil, tomek, michal

Subscribers: