[web] Add to reducer logic handling community id picked in chat
Summary:
In revious diff (D6968) I added to redux state information about the community selected in the Chat tab. This diff adds logic that handles that filed in the reducer, and actions that can be
performed on it. we need to be albe to set the id, and clear it.

Test Plan: Dispatched the new actions programatically, and checked that the redux state changes as planned.

Reviewers: kamil, tomek, michal

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6969