[native] Fix useIsCurrentUserStaff hook
Summary: This hook was introduced in D5498 and incorrectly used `useSelector` from `react-redux` which isn't typed. That meant that the result of this hook was incorrectly typed as `boolean` while in fact it was something like `undefined | null | string | boolean`. The issue wasn't detected because we were simply checking truthiness of the value. In D6949 we started using this value as a query param and converted it to it using `.toString` method, which was failing when the returned value was null / undefined.

Test Plan: After updating the import Flow checks the types correctly. Also logged out and checked if fetching feature flags config works correctly.

Reviewers: atul, kamil, inka, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D6967