[keyserver] update compare-user.js to get missing users by querying identity service
Summary:
We get two lists of users, those in identity but not keyserver and vice versa. We log the users missing from keyserver (this probably means a delete call to identity failed). We delete all ios/android cookies for the users missing from identity to trigger a new login request on these users' devices (in the background).

This triggered request will then populate the identity service with the relevant user data.

Test Plan:
1. while disconnected from the identity service, i registered a new user on ios simulator
2. logged in as this user on web
3. spun up identity service, and when cronjob ran, my new user was added to the test ddb table
4. my web cookie remained in mariadb the whole time, and i was never logged out on either client
5. added a user to ddb manually, and it was logged by keyserver as missing, as expected

Reviewers: ashoat!

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D6998