[native] Replace LayoutAnimation for driving fixed Tooltip animation
Summary:
Driving this separately via `LayoutAnimation` looks janky. This diff makes us drive it using the same `position` that the rest of the animation is driven with.

Depends on D7014

Test Plan: Flow, and testing the reaction popover animation on iOS and Android release builds, including opening the emoji keyboard

Reviewers: ginsu

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D7015