[native] Move NotifPermissionAlertInfo to lib
Summary: We want to show the notif permission modal on web less often and we can reuse NotifPermissionAlertInfo for this. This diff moves it from native redux to the shared redux setup and the base reducer. The next diff will actually use it on web.

Test Plan: Check if redux is correctly updated after showing the alerts on native. Test both a fresh device and when updating. The web was tested with the next diff, and it also correctly updates redux.

Reviewers: inka, tomek

Reviewed By: tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7029