[web] Added displaying 'Labels' to InlineEngagment
Summary: Added the possibility to label the messages to display 'Edited' in the next diffs.

Test Plan: Run the app. Checked if the labels are displayed correctly both with text messages and with images. Checked if Reactions/Threads are also correctly positioned next to the label. Checked if labels in a few messages in a row do not take too much space.

Screenshots: todo

Reviewers: michal, inka, ginsu