[lib] Modify `getClientResponsesSelector` to return `Promise`
Summary: In order to get the `signedIdentityKeysBlob`, we're going to need to `await` `commCoreModule.getUserPublicKey`. As a first step, we refactor `getClientResponsesSelector` so we'll be able to `await` `getUserPublicKey` inside.

Test Plan:
Set breakpoints and ensure that `native`/`web` send responses as expected and `keyserver` continues to get responses as expected. Will also be tested implicitly by subsequent diffs.

Haven't done this yet, will test before landing.

Reviewers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D7034