[desktop] Handle autoUpdate errors
Summary:
We want to handle autoUpdate errors (in particular error about no network connection). If we add a listener to an `error` event electron no longer displays the native alert with the error. Now we can just log it and restart the update schedule.

This replaces D6683 and D6684 as it's a much better solution.

Test Plan: Run the app, turn of internet connection. Check if the error is displayed and the app continues to check for updates.

Reviewers: tomek, kamil

Subscribers: