[keyserver] Give all members permissions to add users to a community
Summary:
Until we have support for roles in the UI, roles are entirely a function of thread type.

This diff updates communities (except GENESIS) so that all members have permissions to add somebody.

It also adds a migration to recalculate permissions for all threads.

Test Plan: I tested this in my local environment:

1. I ran the migration locally
  - It took about 3.6 minutes, but prod has about 32% more threads than my local environments
  - Prod is also likely slower than my MBP. Assuming it's about 50% slower, I'm estimating it will take about 7.12min for this migration to run on prod
  - For that reason, I'll wait until late in the night before deploying it
2. I created a new community locally via `keyserver/src/scripts/create-community.js`
  - I then added a rando test user to it, and confirmed the rando test user could add their friends

Reviewers: ginsu, atul

Subscribers: