[native] Implement `getSignedIdentityKeysBlob(...)` for `native`
Summary:
In this diff we implement `getSignedIdentityKeysBlob(...)` and pass it to `nativeGetClientResponsesSelector...)` so we can call it from within the function created by `getClientResponsesSelector(...)` if we need to handle `serverRequestTypes.SIGNED_IDENTITY_KEYS_BLOB`.

---

Depends on D7034

Test Plan: See if `ClientClientResponse` for `serverRequestTypes.SIGNED_IDENTITY_KEYS_BLOB` is properly constructed and received by `keyserver`.

Reviewers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D7035