[web] Introduce `getSignedIdentityKeysBlobSelector(...)` to generate `getSignedIdentityKeysBlob(...)`
Summary:
Unlike `native`, where in D7035 `getSignedIdentityKeysBlob(...)` makes a JSI call, for `web` we need to construct `signedIdentityKeysBlob` from the Redux `CryptoStore`. We generate `getSignedIdentityKeysBlob(...)` from the new `getSignedIdentityKeysBlobSelector` and pass it to `webGetClientResponsesSelector` where it will be passed to the function generated by `getClientResponsesSelector` (it's a lot of creating a function within a thing that creates a function within a thing that creates a function...)

Now that we have this selector, we can conveniently reuse it in `TraditionalLoginForm()` and `SIWELoginForm()`. I'll make that refactor in a subsequent diff.

---

Depends on D7035

Test Plan:
See if `ClientClientResponse` for `serverRequestTypes.SIGNED_IDENTITY_KEYS_BLOB` is correctly constructed and received by `keyserver`.

Will also be tested in the diff where we consume `getSignedIdentityKeysBlobSelector` in `TraditionalLoginForm()` and `SIWELoginForm()`.

Reviewers: ashoat, tomek