[keyserver] Introduce `isCookieMissingSignedIdentityKeysBlob`
Summary:
Checks if there's an existing row in `cookies` table with `id=cookieID` with a `NULL` `signed_identity_keys` column.

We're going to use this check in `handleInitialClientSocketMessage` to determine whether or not to send `SIGNED_IDENTITY_KEYS_BLOB` request to `client` from `keyserver`.

---

Depends on D7049

Test Plan:
Called function from `handleInitialClientSocketMessage`:

```lang=typescript
    if (viewer.cookieID) {
      isCookieMissingSignedIdentityKeysBlob(viewer.cookieID);
    }
```

and set breakpoint within `isCookieMissingSignedIdentityKeysBlob` to check whether `queryResult` was as expected when `signed_identity_keys` was `NULL` and a valid stringified SignedIdentityKeysBlob:

Reviewers:
ashoat, tomek