[web] Move logic for fetching ThreadInfo for possibly pending chat out of ChatMessageListContainer component
Summary:
Issue: https://linear.app/comm/issue/ENG-2740/add-navigation-state-info-to-the-top-bar
The new designs call for the navigation state info to be moved from message list header, to main content's header. The component that is used for displaying the nav state info ('ThreadTopBar') needs the ThreadInfo of the open chat, or
a special ThreadInfo object when a chat is being created. The excrated logic is used to obtain that exact ThreadInfo. This will allow me to use it in the `Topbar` component, where I'll need to display the nav state info.

Test Plan: Run web app, check that the `ThreadTopBar` component displys correctly in the message list header

Reviewers: michal, kamil, kuba

Reviewed By: kamil

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D6741