Use initial notifications from queue rather than from launch options
Summary: This differential is a fix for the iOS part of this issue: https://linear.app/comm/issue/ENG-3269/app-erroneously-opens-to-chats-at-random-times. Initially when app was entering the foreground in respone to tapping the notif we were querying
initial launch options for notification that launched the app (no the one that was tapped, but the one that booted the app). In case there was any we would navigate user to the thread corresponding to this notification. This resulted in erroneous navigation since the thread correpsonding to
the lastest notif that foregrounded (not booted!) the app might had been different.
Test Plan: Build the app and check that tapping notification always navigates us to the correct thread.

Reviewers: atul, bartek, tomek

Subscribers: