[native] Move 'InlineEngagement' styling from 'ComposedMessage' to component
Summary: Move 'InlineEngagement' styling from 'ComposedMessage' to 'InlineEngagement' component to make it concise with the web version. It makes it also easier to add message labels to the code, so they were in the 'InlineEngagement' component, like on the web.

Test Plan: Checked if everything works after the changes. Nothing visible should change. Compared how reactions and threads are rendered in iOS and Android simulators with the old
version.

Reviewers: michal, inka, ginsu

Subscribers: