[keyserver] Introduce `setCookieSignedIdentityKeysBlob`
Summary:
Introduce `setCookieSignedIdentityKeysBlob` to update `signed_identity_keys` column of row in `cookies` table with `id`=`cookieID`.

This is to update serverDB when client `SIGNED_IDENTITY_KEYS_BLOB` response is received.

---

Depends on D7048

Test Plan: Ran query of that form manually in query console and it worked as expected. Will be tested implicitly by subsequent diffs in this stack.

Reviewers: ashoat

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D7049