[native] render avatars in chat screen
Summary: Rendered avatars in the chat screen. Avatars only render for users who are not the viewer and at the end of a meassage cluster. When the message gets swiped the avatar moves with the message

Test Plan: Please watch the demo videos to see all the edge cases I tested

Reviewers:

Subscribers: