[lib][keyserver] Update reaction permission checks to check REACT_TO_MESSAGE
Summary:
In the last diff, I introduced a new `threadPermissions` constant for reactions.

In this diff, I update permission checks on the keyserver and client to check this new constant.

Depends on D7073

Test Plan: I tested this in combination with the next diff by:

1. Running the migration by restarting the `keyserver` after introducing it and making sure it runs successfully
2. Testing reacting as a member (non-admin) to a message in a `COMMUNITY_ANNOUNCEMENT_ROOT` community root. I don't have permission to send messages, but I can react on both native and web.
I didn't need to log out / back in on native, which indicates the the `updatesTypes.UPDATE_THREAD` updates were propagated correctly by the migration

Reviewers: ginsu, kuba, rohan

Subscribers: